Add snap channel support
Add a new configuration option 'channel' to support deployment
of vault from channels other than stable.

This superceeds the layer option which is dropped as part of this
review.

Channel config option changes are honoured, but vault may not be
restarted depending on whether the local vault service has been
unsealed.

Change-Id: I845cd8db3a7e80155f2b485bb3f07934476996f2