Add support for new secrets:vault-kv interface
Provide implementation for the provides endpoint of the vault-kv
typed interface.

This interface type will setup a KV secret backend and create an
approle and policy to allow remote services to access the backend
using the Vault API.

Backends may be shared between remote units, allowing any unit to
access any value in the backend, or may be isolated between units
based on path suffixing using the units hostname so that stored
secrets are not visible between units of a deployment.

Change-Id: Id8fa1cbe33feccc9c2f06a61db22453d7830730d