Backport Zaza reconfiguration
By splitting up Zaza from the openstack-charms specific tests,
we can more easily guarantee test stability, while not harming
out ability to improve on the testing library during a cycle.

This change backports the zaza namespace change landing, and
adds a stable branch of the zaza-openstack-charms project to
the test-dependencies.

Change-Id: I0c3be578938e1ef8b2752d5cc6883043bda0d110