mysql: configure maximum connection lifetime
Ensure that any server side termination of MySQL connections is
handled well by configuring a maximum connection lifetime of
3600 seconds. This is inline with defaults in OpenStack/oslo.db
and is typically equal to or less that the wait_timeout configured
in MySQL or Percona XtraDB Cluster.  The default is '0' which
means connections will be used forever.

Longer term solution to this issue is to pass the configured
wait_timeout in MySQL on the shared-db relation so clients can
be configured consistently with the MySQL server.

Change-Id: Icd76caae0ff56fc5a27ea08399cbc6af59852897
Related-Bug: 1841063