Ensure we clear the ca.ready flag when we generate a CA
Previously, configuring any kind of CA, and then generating a new
CA with CSR will cause the charm to have hook errors, as well
as an error in the action to generate a CSR as a result of
expecting a fully setup CA elsewhere in the charm. This change
adds removal of that ca.ready flag from the charm when a CSR
(and accompanying CA) are generated.

Change-Id: I802f31c853df4a69f6a4c1529e2acb1543e21831
Closes-Bug: #1866150