Pin tox to < 4.0.0
Tox 4.0.0 was recently released and it has several breaking changes.
We pin to < 4.0.0 here. We are planning to move forward only on the
master charm branches.

Tox is also pinned to < 4.0.0 for stable branches in upstream
openstack-zuul-jobs as well as in zosci-config. However, the
requires= section in the charm's tox.ini file ends up installing
tox>4, wiping out the zuul-pinned tox<4 that was already installed
installed. This patch fixes that.

Additional changes included in this patch:
- Add bindep.txt to install standard binary dependencies.
- Switch charm-tools to 2.8.4 to remove ruamel requirement that
  doesn't work on Python 3.10 (required for pep8/cover).
- Pin zaza and zaza-openstack-tests to stable/yoga branch as
  stable/21.10 is no longer supported.
- Pin urllib3 in test-requirements.txt for py35 unit tests

Related-Bug: #2002788
Change-Id: Ib60536bae9fa69ec9966407dd352e521528cb6f1