Add docs key and point at Discourse
Add the 'docs' key and point it at a Discourse topic
previously populated with the charm's README contents.

When the new charm revision is released to the Charmhub,
this Discourse-based content will be displayed there. In
the absense of the this new key, the Charmhub's default
behaviour is to display the value of the charm's
'description' key.

Change-Id: I51e3ce5347f2036165429145075e15c9801a26af
(cherry picked from commit ecd78eb53c1c84e127e0b01cbbf258a616db665e)