Fix NRPE check return error after snap upgrade
A recent update to snapd changed the response snapd returns when
requesting the version of the Vault snap. This is due to the snapd
API talking http over a socket interface and the code in question
not actually understanding http protocol. Rather than add additional
dependencies on the machine for this check to run when invoked by
nagios invocations, the code is rewritten to parse the version of
the installed vault snap from the output of the snap info command.

Closes-Bug: 2060173
Change-Id: I1d6b9d05b960e7723fab30594dcb73769066112c
(cherry picked from commit f43f7e2d774e449ba51823af5fc2f82e2ff13890)